backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
openai
sqlmodel
requests
elevenlabs
reportlab
stripe
